api/requirements.txt
supabase-py==2.4.0
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
